% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cnv_segment_annotation}
\alias{cnv_segment_annotation}
\title{Function that annotates CNV segment files (FACETS)}
\usage{
cnv_segment_annotation(cnv_file, format = "facets")
}
\arguments{
\item{cnv_file}{from FACETS}

\item{format}{}
}
\value{
cnv_data
}
\description{
Function that annotates CNV segment files (FACETS)
}

