% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_project_reports}
\alias{generate_project_reports}
\title{Function that generate cancer genome reports for a project cohort}
\usage{
generate_project_reports(project_directory, strelka_snv_vcf_gz,
  mutect_snv_vcf_gz, strelka_indel_vcf_gz = NULL,
  project_sample_stats = NULL, signatures.limit = 10,
  strong_calls_only = FALSE, project_name = "NCGC_tumor_exome",
  write_project_files = FALSE)
}
\arguments{
\item{project_directory}{name of project directory}

\item{strelka_snv_vcf_gz}{name of bgzipped VCF file with annotated Strelka SNVs}

\item{mutect_snv_vcf_gz}{name of bgzipped VCF file with annotated MuTect SNVs}

\item{strelka_indel_vcf_gz}{name of bgzipped VCF file with annotated Strelka indels}

\item{project_sample_stats}{sample_stats data frame}

\item{signatures.limit}{Number of signatures to limit mutational signature analysis}

\item{strong_calls_only}{Logical indicating if report is based on strong calls only or all calls}

\item{project_name}{Name of project}

\item{write_project_files}{Logical indicating if global, cohort-levels files are to be printed}
}
\value{
p
}
\description{
Function that generate cancer genome reports for a project cohort
}

