% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_report_data}
\alias{generate_report_data}
\title{Function that generates multiple reports}
\usage{
generate_report_data(sample_calls, sample_id = NULL,
  minimum_n_signature_analysis = 50, signatures.limit = 6)
}
\arguments{
\item{sample_calls}{data frame with list of variant calls}

\item{sample_id}{sample identifier}
}
\value{
report_data data frame with all report elements
}
\description{
Function that generates multiple reports
}

