% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_tier_tsv}
\alias{generate_tier_tsv}
\title{Function that generates dense and tiered annotated variant datasets}
\usage{
generate_tier_tsv(tier1_variants, tier2_variants, tier3_variants,
  tier4_variants, tier5_variants, sample_id = "test")
}
\arguments{
\item{sample_id}{Sample identifier}

\item{report_data}{report data structure generated from pcgr}
}
\value{
tsv_variants data frame with tier-annotated list of variants for tab-separated output
}
\description{
Function that generates dense and tiered annotated variant datasets
}

