% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_clinical_associations_civic}
\alias{get_clinical_associations_civic}
\title{Function that appends clinical annotations for somatic cancer variants}
\usage{
get_clinical_associations_civic(vcf_data_df, mapping = "exact",
  ncgc = FALSE)
}
\arguments{
\item{mapping}{- one of 'exact' (allele-specific) or 'approximate' (codon or exon-level biomarkers)}

\item{ncgc}{- logical indicating whether NCGC-specific tags are to be appended}
}
\value{
vcf_data_df
}
\description{
Function that appends clinical annotations for somatic cancer variants
}

