% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_call_statistics}
\alias{plot_call_statistics}
\title{Function that plots call numbers}
\usage{
plot_call_statistics(sample_calls, title = "Title")
}
\arguments{
\item{sample_calls}{data frame with sample variants}

\item{title}{Title for plot}
}
\value{
p
}
\description{
Function that plots call numbers
}

