% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{signature_contributions_single_sample}
\alias{signature_contributions_single_sample}
\title{Function that retrieves relative estimates of known somatic signatures from a single tumor}
\usage{
signature_contributions_single_sample(mut_data, sample_name,
  signatures_limit = 10)
}
\arguments{
\item{signatures.limit}{}
}
\description{
Function that retrieves relative estimates of known somatic signatures from a single tumor
}

