% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_project_reports}
\alias{generate_project_reports}
\title{Function that plots call numbers}
\usage{
generate_project_reports(project_directory, strelka_snv_vcf_gz,
  mutect_snv_vcf_gz, strelka_indel_vcf_gz = NULL,
  project_sample_stats = NULL, signatures.limit = 10)
}
\arguments{
\item{project_directory}{name of roject directory}

\item{strelka_snv_vcf_gz}{name of bgzipped VCF file with annotated Strelka SNVs}

\item{mutect_snv_vcf_gz}{name of bgzipped VCF file with annotated MuTect SNVs}

\item{strelka_indel_vcf_gz}{name of bgzipped VCF file with annotated Strelka indels}

\item{project_sample_stats}{sample_stats data frame}

\item{signatures.limit}{Number of signatures to limit mutational signature analysis}
}
\value{
p
}
\description{
Function that plots call numbers
}

