% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_tier_tsv}
\alias{generate_tier_tsv}
\title{Function that generates tier datasets}
\usage{
generate_tier_tsv(report_data)
}
\arguments{
\item{report_data}{}
}
\value{
tsv_variants data frame with tier-annotated list of variants for tab-separated output
}
\description{
Function that generates tier datasets
}

